+++ test/CompTest.hs
@@ -1,9 +1,5 @@
-{-# OPTIONS_GHC
-    -fno-warn-orphans
-#-}
-{-# OPTIONS_GHC -ddump-tc-trace -ddump-to-file
-#-}
-{-# OPTIONS_GHC -fplugin Control.Monad.CheckedExcept.Plugin -fplugin-opt Control.Monad.CheckedExcept.Plugin:verbose  #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+{-# OPTIONS_GHC -fplugin Control.Monad.CheckedExcept.Plugin #-}
 
 {-# LANGUAGE
     TypeApplications
@@ -16,64 +12,59 @@
 
 module CompTest where
 
--- import Test.Tasty.HUnit
 import Control.Monad.CheckedExcept
 import Control.Monad.Trans.Class (lift)
 import qualified Control.Monad.CheckedExcept.QualifiedDo as CheckedExcept
 
--- badTestCE1 :: CheckedExceptT '[Int] IO ()
--- badTestCE1 = CheckedExcept.do
---   lift $ putStrLn "bad 1"
---   throwCheckedException 'c' -- this shouldn't work
-
+-- Single-exception helpers use plain @do@ (signature fixes @es@). Only the
+-- composing @testCE@ block needs @CheckedExcept.do@ to union exception sets.
 testCE1 :: CheckedExceptT '[()] IO ()
-testCE1 = CheckedExcept.do
+testCE1 = do
   lift $ putStrLn "1"
-  lift $ pure ()
   pure ()
 
 testCE2 :: CheckedExceptT '[Int] IO ()
-testCE2 = CheckedExcept.do
+testCE2 = do
   lift $ putStrLn "2"
   throwCheckedException (1 :: Int)
   pure ()
 
 testCE3 :: CheckedExceptT '[Bool] IO ()
-testCE3 = CheckedExcept.do
+testCE3 = do
   lift $ putStrLn "3"
   throwCheckedException False
   pure ()
 
 testCE4 :: CheckedExceptT '[String] IO ()
-testCE4 = CheckedExcept.do
+testCE4 = do
   lift $ putStrLn "4"
   throwCheckedException "err"
   pure ()
 
 testCE5 :: CheckedExceptT '[Char] IO ()
-testCE5 = CheckedExcept.do
+testCE5 = do
   lift $ putStrLn "5"
   throwCheckedException 'c'
   pure ()
 
 testCE :: CheckedExceptStack ()
-testCE = CheckedExceptStack $ CheckedExcept.do
-  () <- testCE1
-  () <- testCE2
-  () <- testCE3
-  () <- testCE4
-  -- () <- testCE5 -- doesn't compile
-  pure ()
+testCE =
+  CheckedExceptStack $
+    (CheckedExcept.do
+      () <- testCE1
+      () <- testCE2
+      () <- testCE3
+      () <- testCE4
+      -- () <- testCE5 -- doesn't compile
+      pure () :: CheckedExceptT TestExceptions IO ())
 
 test :: CheckedExcept TestExceptions () -> IO ()
 test ce = case runCheckedExcept ce of
   Left e -> do
     applyAll (putStrLn . encodeException) e
-    -- or
     withOneOf @() e $ \() -> putStrLn "()"
     withOneOf @Int e $ \n -> print $ n + 1
     withOneOf @Bool e $ \_ -> pure ()
-    -- or
     caseException e
       (  (\() -> putStrLn "()")
       <: (\n -> print $ n + 1)
@@ -86,16 +77,6 @@ test ce = case runCheckedExcept ce of
       <: CaseAny (\x -> putStrLn $ encodeException x)
       )
   Right () -> putStrLn "Right"
-
-{-
--- doens't compile
-test2 :: CheckedExcept '[] () -> IO ()
-test2 ce =
-  case runCheckedExcept ce of
-    Left (OneOf e) -> do
-      caseException (Proxy @'[]) e Nil
-    Right () -> putStrLn "Right"
--}
 
 type TestExceptions = '[(), Int, Bool, String]